RecordsController: change record creation workflow, add update method
In order to have consistent record data for the UI, the workflow to
create a new record has been changed. When going to create a new record,
call prepare() at first to instantiate a new record object before adding
it to the database. Then use add() to add the new data to the database.
To finish the current recording, use finish() as before.

The new update() method takes a pointer to a record object and updates
the record's data in the database.