Add delete option to single record page.
When using the delete option on the single record page, the behavior
depends on where the page was opened from. If it was opened from a list,
set comingFromList to true will show the delete remorse on the list
item. If not coming from a list, the remorse popup will be shown on the
page and the page will be popped back after the remorse popup was
executed.